+++ Spyen/src/Renderer/VertexArray.cpp
@@ -60,7 +60,7 @@ namespace Spyen {
 					ShaderDataTypeToOpenGLBaseType(element.Type),
 					element.Normalized ? GL_TRUE : GL_FALSE,
 					layout.GetStride(),
-					reinterpret_cast<const void*>(element.Offset));
+					reinterpret_cast<const void*>(static_cast<uintptr_t>(element.Offset)));
 				m_VertexBufferIndex++;
 				break;
 			}
@@ -74,7 +74,7 @@ namespace Spyen {
 						ShaderDataTypeToOpenGLBaseType(element.Type),
 						element.Normalized ? GL_TRUE : GL_FALSE,
 						layout.GetStride(),
-						reinterpret_cast<const void*>(element.Offset + sizeof(float) * count * i));
+						reinterpret_cast<const void*>(static_cast<uintptr_t>(element.Offset + sizeof(float) * count * i)));
 					glVertexAttribDivisor(m_VertexBufferIndex, 1);
 					m_VertexBufferIndex++;
 				}
@@ -91,7 +91,7 @@ namespace Spyen {
 					element.GetComponentCount(),
 					ShaderDataTypeToOpenGLBaseType(element.Type),
 					layout.GetStride(),
-					reinterpret_cast<const void*>(element.Offset));
+					reinterpret_cast<const void*>(static_cast<uintptr_t>(element.Offset)));
 				m_VertexBufferIndex++;
 				break;
 			}
