+++ Spyen/src/Renderer/VertexBuffer.h
@@ -6,12 +6,12 @@ namespace Spyen {
 	class VertexBuffer
 	{
 	public:
-		VertexBuffer(const void* data, uint32_t size);
+		VertexBuffer(const void* data, size_t size);
 		~VertexBuffer();
 
 		void Bind() const;
 		void Unbind() const;
-		void SetData(const void* data, uint32_t size);
+		void SetData(const void* data, size_t size);
 		void SetLayout(const BufferLayout& layout) { m_Layout = layout; };
 		const BufferLayout& GetLayout() const { return m_Layout; };
 		uint32_t& GetRendererID() { return m_RendererID; }
